Now included -
Formatting for EM as specified in em_vector.

Note: performance on build_explicit_models is crappy so that's going to need some more work, but it should work for testing EM for the moment.

Also cleaned up readFiles_test, not quite ready to be deleted yet, but it will be eventually.